delete main.py and rename main_new.py to main.py
